fix: 클로드 리뷰 반영
